requirements.txt
Django==5.2.6
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
Pillow==10.4.0
qrcode==7.4.2
plotly==5.22.0
pandas==2.2.2
django-extensions==3.2.3
django-import-export==4.1.1
django-tables2==2.7.0
django-filter==24.3
